requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.0
APScheduler==3.10.4
TgCrypto==1.2.5